+++ voidstar/algos_3d/sphere_conti_rainbow.cc
@@ -6,15 +6,15 @@ class Algo3DSphereContiRainbow : public Algo3DSphere {
   virtual ~Algo3DSphereContiRainbow() {}
 
   virtual bool apply(Floats& vertices, Floats& colors, VertIndices& indices,
-                     size_t width, size_t height, size_t depth) final {
+                     u32 width, u32 height, u32 depth) final {
     make_vertices(vertices, width, height, depth);
 
-    size_t size;
+    u32 size;
     const u8* data = loadDataRange(size);
     u8 x = data[0];
     u8 y = data[1];
 
-    for (size_t i = 2; i < size; ++i) {
+    for (u32 i = 2; i < size; ++i) {
       u8 z = data[i];
       Index id = x + y * height + z * depth * height;
       Index idx = 3 * id;
