+++ voidstar/algos_3d/sphere_full.cc
@@ -6,10 +6,10 @@ class Algo3DSphereFull : public Algo3DSphere {
   virtual ~Algo3DSphereFull() {}
 
   virtual bool apply(Floats& vertices, Floats& colors, VertIndices& selected,
-                     size_t width, size_t height, size_t depth) final {
+                     u32 width, u32 height, u32 depth) final {
     auto nb_verts = make_vertices(vertices, width, height, depth);
 
-    for (size_t i = 0; i < nb_verts; ++i) {
+    for (u32 i = 0; i < nb_verts; ++i) {
       colors[3 * i + 0] = 1.0f;
       colors[3 * i + 1] = 1.0f;
       colors[3 * i + 2] = 1.0f;
